Move dialogs into components
